private KEY/requirements.txt
streamlit 
langchain
transformers
youtube-transcript-api
spacy
tqdm
pydantic
python-dotenv
